Fix heap debug assertion failure in Graph::shortestPath(). Use size_t in Polygon2D::insertPoint.

+++ rpg/utils.h
@@ -186,7 +186,7 @@ public:
     virtual void addPoint(Vector2D point) {
         this->points.push_back(point);
     }
-    virtual void insertPoint(int indx, Vector2D point);
+    virtual void insertPoint(size_t indx, Vector2D point);
     void setSource(void *source) {
         this->source = source;
     }
